Updates to Zig v0.14.0 and adds new features
Updated to now support zig version 0.14.0.

The Managed and Unmanaged version of the RBTree implementation now
support constructing a red-black tree from an iterator into a sorted
list, or a slice of a sorted list.

This feature is presently in alpha and will be properly checked over the
next few days.