Updates the code foe version 0.3.0
All functions of the form `*Context` in types created by `RBTreeUnmanaged` have  been renamed as `*WithContext`.
This is to reduce ambiguity when calling functions like `getContext`.

Added some new functions `moveNodeByCopy` and `swapNodeStorage` to `RBTreeImplementation`.
These function enable some more advanced usage.

Updated the implementation of `initFromSortedKVIterator` in `RBTreeUnmanaged` so that it no longer uses recursion.
Moreover, this function now issues a compile error if the input iterator does not match the desired iterator interface.

Added function `isIterator` to `rbtree.meta` in order to check if a given value is an iterator.

Updates the default sort in `rbtree.meta.order` so that it now supports pointer dereferences with a comptime limit on
the number of nested pointer dereferences.
There is also a new function `rbtree.defaultOrderGeneric` which also takes in a pointer depth.