Adds a test for moveNodeByCopyWithCache and adds links to readme
The previous two tagged version of this library were corrections to some
issues found in the code which were not discovered due to lack of proper
testing. After realising that there were no tests for the function
moveNodeByCopyWithCache, I decided to add one. Fortunatly, there were no
errors in the impementation of this function.

This update also adds some new alternative links to the readme.