Removes function used in tests from the documentation
In the previous version, there were two `next` functions mentioned in
the documentation for the `meta` namespace. These functions only
appeared in a test and were not accessible outside of this test. This
update simply changes the visibility of these functions so that they no
longer appear in the documentation.